+++ etils/epy/__init__.py
@@ -20,6 +20,7 @@ import sys
 
 from etils.epy.backports import cached_property
 from etils.epy.contextlib import ContextManager
+from etils.epy.env_utils import is_ipython_terminal
 from etils.epy.env_utils import is_notebook
 from etils.epy.itertools import groupby
 from etils.epy.itertools import splitby