+++ etils/epy/env_utils.py
@@ -21,10 +21,20 @@ import sys
 def is_notebook() -> bool:
   """Returns True if running in a notebook (Colab, Jupyter) environment."""
   # Use sys.module as we do not want to trigger an import (slow)
-  IPython = sys.modules.get('IPython')  # pylint: disable=invalid-name
   # Check whether we're not running in a IPython terminal
-  if IPython:
+  if IPython := sys.modules.get('IPython'):  # pylint: disable=invalid-name
     get_ipython_result = IPython.get_ipython()
     if get_ipython_result and 'IPKernelApp' in get_ipython_result.config:
       return True
   return False
+
+
+def is_ipython_terminal() -> bool:
+  """Returns True if running in a IPython terminal environment."""
+  if IPython := sys.modules.get('IPython'):  # pylint: disable=invalid-name
+    get_ipython_result = IPython.get_ipython()
+    if (
+        get_ipython_result
+        and type(get_ipython_result).__name__ == 'TerminalInteractiveShell'
+    ):
+      return True